chore(types): seperate tag/author/category types
