+++ types/authors.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod'
+
+const AuthorSchema = z.object({
+  id: z.number().optional(),
+  name: z.string().nullable(),
+  link: z.string().nullable(),
+  image: z.string().nullable()
+})
+
+export type Author = z.infer<typeof AuthorSchema>
