+++ types/tags.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod'
+
+const TagSchema = z.object({
+  id: z.number().optional(),
+  name: z.string()
+})
+
+export type Tag = z.infer<typeof TagSchema>
